+++ quicktill/test_models.py
@@ -194,6 +194,47 @@ class ModelTest(unittest.TestCase):
         with self.assertRaises(IntegrityError):
             self.s.commit()
 
+    def test_barcode_useful_constraint(self):
+        stockline, plu = self.template_stockline_and_plu_setup()
+        self.s.add(models.Barcode(barcode='123456'))
+        with self.assertRaises(IntegrityError):
+            self.s.commit()
+
+    def test_barcode_unambiguous_constraint(self):
+        stockline, plu = self.template_stockline_and_plu_setup()
+        self.s.add(models.Barcode(barcode='12121', stockline=stockline,
+                                  plu=plu))
+        with self.assertRaises(IntegrityError):
+            self.s.commit()
+
+    def test_barcode_stockline_cascade(self):
+        stockline, plu = self.template_stockline_and_plu_setup()
+        barcode = models.Barcode(barcode='123456', stockline=stockline)
+        self.s.add(barcode)
+        self.s.commit()
+        self.assertEqual(stockline.barcodes, [barcode])
+        self.s.delete(stockline)
+        self.s.commit()
+
+    def test_barcode_plu_cascade(self):
+        stockline, plu = self.template_stockline_and_plu_setup()
+        barcode = models.Barcode(barcode='123456', plu=plu)
+        self.s.add(barcode)
+        self.s.commit()
+        self.assertEqual(plu.barcodes, [barcode])
+        self.s.delete(plu)
+        self.s.commit()
+
+    def test_barcode_delete(self):
+        stockline, plu = self.template_stockline_and_plu_setup()
+        barcode = models.Barcode(barcode='123456', plu=plu)
+        self.s.add(barcode)
+        self.s.commit()
+        self.s.delete(barcode)
+        self.s.commit()
+        self.s.refresh(plu)
+        self.assertEqual(plu.barcodes, [])
+
     def test_transline_void(self):
         self.template_setup()
         session = models.Session(datetime.date.today())